Kernel - Finally fixed that corruption bug (description follows)
- Also fixed some logging bugs (Heap dump and VFS)
- Fixed a bug in the AxWin2 shell include statement

Bug Description:
The last non-idle thread to be put to sleep didn't get the ->CurCPU field
reset to -1. This probably caused the scheduler to think that the process
was still in use, and hence it never got re-scheduled. The other VTs did
not suffer this problem because they were not the last to sleep.